+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Enhancements
   * [Docker] Add command `azk docker`
   * [Vm] Adding ssh keys generator, and uploads the key to start vm.
+  * [Generators] Now Generators and rules is a extension of the UI.
 
 * Bug
   * [Cli] Fix generate manifest in blank project dir.