+++ CHANGELOG.md
@@ -6,6 +6,11 @@
   * [Docker] Add command `azk docker`
   * [Vm] Adding ssh keys generator, and uploads the key to start vm.
   * [Generators] Now Generators and rules is a extension of the UI.
+
+## v0.3.1 (2014-07-11)
+
+* Enhancements
+  * [Manifest] Support `env` to use environment variables in DSL.
   * [Generators] Adding base initializar for ruby projects.
 
 * Bug